Chars that are used to store numbers... need to be cast before they are printed.

+++ rvm/src/org/jikesrvm/osr/OSR_VariableElement.java
@@ -209,7 +209,7 @@ public class OSR_VariableElement implements OSR_Constants {
     } else {
       buf.append('S');
     }
-    buf.append(num);
+    buf.append((int)num);
     buf.append(",");
 
     char t = 'V';
